First commit after getting basics up and running
Should have made a repo to begin with, but got caught up in coding.
Synth works okay, especially with signal.square or signal.saw, but the
buffer appears to starve a little with np.sin.
Need to attenuate waves based on frequency, since amplitude is not the
only thing affecting loudness.
Need to make buttons work better.
Need to make a GUI.