More notes added
The synth now has 16 notes available, with keys from s-æ.
The note pattern from s-h is a copy of the note pattern from j-ø. This
means it does not fit into the same pattern. The reason for this is that
j-ø are intended as the main keys to play notes with, whereas the keys
farther left are mostly intended to help play chord inversions easily.